pyproject.toml
[project]
name = "faq-agent"
version = "0.1.0"
description = "Faq Agent built in Google AI ecosystem"
authors = [
    { name = "AnishNavalgund", email = "anishk.navalgund@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.11"

dependencies = [
    "google-cloud-aiplatform>=1.38.0,<2.0.0",
    "google-cloud-storage>=2.12.0,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
